+++ src/main/scala/com/kinetica/spark/datasourcev2/KineticaDataReader.scala
@@ -1,13 +1,15 @@
-package com.kinetica.spark.datasourcev2
+package com.kinetica.spark.datasourcev2;
 
-import org.apache.spark.sql.sources.v2.reader.DataReader
-import org.apache.spark.sql.catalyst.expressions.Expression
-import org.apache.spark.sql.Row
-import org.apache.spark.sql.types.StructType
-import com.kinetica.spark.LoaderParams
-import com.kinetica.spark.egressutil.KineticaUtils
-import com.kinetica.spark.util.ConfigurationConstants._
-import com.typesafe.scalalogging.Logger
+import org.apache.spark.sql.sources.v2.reader.DataReader;
+import org.apache.spark.sql.catalyst.expressions.Expression;
+import org.apache.spark.sql.Row;
+import org.apache.spark.sql.types.StructType;
+import com.kinetica.spark.LoaderParams;
+import com.kinetica.spark.egressutil.KineticaEgressUtilsNativeClient;
+import com.kinetica.spark.egressutil.KineticaEgressUtilsJdbc;
+import com.kinetica.spark.util.ConfigurationConstants._;
+// import com.typesafe.scalalogging.Logger;
+import com.typesafe.scalalogging.LazyLogging;
 
     
 class KineticaDataReader (
@@ -15,9 +17,10 @@ class KineticaDataReader (
     val tableSchema: StructType,
     val pushedCatalystFilters: Array[Expression],
     val requiredSchema: StructType)
-    extends DataReader[Row] {
+    extends DataReader[Row]
+    with LazyLogging {
  
-    val logger = Logger("KineticaDataReader");
+    // val logger = Logger("KineticaDataReader");
   
     val requiredColumns = requiredSchema.fieldNames;
 
@@ -29,18 +32,42 @@ class KineticaDataReader (
     // if ( requiredColumns.isEmpty ) {
     // }
 
+
     val gpudbConn = conf.getGpudb;
     val tableName = conf.getTablename;
     
     // Get the table size
-    val tableSize = KineticaUtils.getKineticaTableSize( gpudbConn, tableName );
+    val tableSize = KineticaEgressUtilsNativeClient.getKineticaTableSize( gpudbConn, tableName );
+    logger.debug("KineticaDataReader: got table size {}", tableSize);
      
     // Read the table's rows
-    val batchSize = 10000
-    val myRows = KineticaUtils.getRowsFromKinetica( gpudbConn, tableName,
-                                                    requiredColumns, requiredSchema,
-                                                    0, tableSize.toInt, batchSize );
-  
+    val batchSize = 10000;
+
+    // Retrieve the records
+    val myRows: Iterator[Row] = {
+
+        if ( pushedCatalystFilters.isEmpty ) {
+            logger.debug("KineticaDataReader: No filters given; use the native client API");
+            // Use the fastest path via the native Java API to get the records
+            // since no filters are used
+            KineticaEgressUtilsNativeClient.getRowsFromKinetica( gpudbConn, tableName,
+                                                                 requiredColumns, requiredSchema,
+                                                                 0, tableSize.toInt, batchSize );            
+        } else {
+            logger.debug("KineticaDataReader: Filters ARE given; use the JDBC connector");
+            // Use the JDBC connector to get the records
+            val queryStr = buildTableQuery(table, requiredColumns, pushedCatalystFilters)
+
+            val conn = com.kinetica.spark.egressutil.KineticaJdbcUtils.getConnector(url, conf)();
+            val stmt = conn.prepareStatement( queryStr );
+            val rs = stmt.executeQuery;
+        
+            val internalRows = KineticaEgressUtilsJdbc.resultSetToSparkInternalRows( rs, requiredSchema );
+            val encoder = org.apache.spark.sql.catalyst.encoders.RowEncoder.apply( requiredSchema ).resolveAndBind();
+            internalRows.map( encoder.fromRow );
+        }
+    }
+
   
     override def next(): Boolean = {
         myRows.hasNext
@@ -52,7 +79,8 @@ class KineticaDataReader (
 
     override def close(): Unit = {
     }
-  
+
+    
     private def buildTableQuery(
         table: String,
         columns: Array[String],
@@ -84,7 +112,7 @@ class KineticaDataReader (
             // interpolation in scala; the following is correct, though ugly
             s"""SELECT $colStrBuilder FROM "$table" $whereClause"""
         }
-        logger.info("External Table Query: " + baseQuery)
+        logger.info("Query for retrieving records: " + baseQuery)
         baseQuery.toString()
     }   // end buildTableQuery